requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-genai==0.2.2
pillow==10.0.1
python-dotenv==1.0.0
werkzeug==2.3.7